log.py became more usable
There is no difference between the module and it's functions.
So you can use log.warn() or log.warn.debug() or whatever.

This greatly makes it more easy to just pass the wanted function
around, but you are also able to switch to some other function
on the fly, without need to pass the logging module at all.